:hammer: Fixed type index declaration
https://github.com/Gjacquenot/test_clang/runs/7746289257?check_suite_focus=true

[615/908] Building CXX object grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
FAILED: grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o
/usr/bin/clang++ -DGTEST_HAS_POSIX_RE=0 -I. -I/opt/share/code/grpc/inc -I/opt/share/code/ssc -I/opt/share/code/exceptions/inc -I/opt/share/code/external_data_structures/inc -I/opt/share/code/core/inc -I/opt/share/code/environment_models/inc -I/opt/share/code/hdb_interpolators/inc -I/opt/share/code/mesh/inc -isystem /opt/eigen -isystem /opt/boost/include -isystem /opt/grpc/include -isystem grpc -isystem /opt/share/code/yaml-cpp/src -isystem /opt/share/code/yaml-cpp/include -isystem /opt/share/code/yaml-cpp/include/yaml-cpp -Werror -pedantic -pedantic-errors -Wextra -Wall -Wunused-function -Wunused-label -Wunused-parameter -Wunused-value -Wunused-variable -fno-common -Wformat=2 -Winit-self -Wpacked -Wpointer-arith -Wmissing-declarations -Wmissing-format-attribute -Wsign-compare -Wstrict-aliasing=2 -Wundef -Wconversion -Wno-misleading-indentation -Woverloaded-virtual -Weffc++ -Wwrite-strings -Wfatal-errors -Wno-deprecated -Wvariadic-macros -fPIC -std=gnu++11 -O3 -DNDEBUG -MD -MT grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o -MF grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o.d -o grpc/CMakeFiles/grpc.dir/src/SurfaceElevationFromGRPC.cpp.o -c /opt/share/code/grpc/src/SurfaceElevationFromGRPC.cpp
/opt/share/code/grpc/src/SurfaceElevationFromGRPC.cpp:60:99: fatal error: implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]
                throw_invalid_size_error(rpc_method, "XYZTGrid", "y", response.x_size(), response.y_size());
                ~~~~~~~~~~~~~~~~~~~~~~~~                                                 ~~~~~~~~~^~~~~~~~
1 error generated.

+++ code/grpc/src/SurfaceElevationFromGRPC.cpp
@@ -38,7 +38,7 @@ class SurfaceElevationFromGRPC::Impl
             throw_if_invalid_status("set_parameters", status);
         }
 
-        void throw_invalid_size_error(const std::string& rpc_method, const std::string& grpc_type, const std::string& field, const size_t expected_size, const size_t actual_size) const
+        void throw_invalid_size_error(const std::string& rpc_method, const std::string& grpc_type, const std::string& field, const int expected_size, const int actual_size) const
         {
             std::stringstream ss;
             ss << "An error occurred when using the distant wave model defined via gRPC (method '"
